CowREFCNT: new sv_cowrefcnt, shared with sv_refcnt
8+24 bit for the U32 refcnt in HEAD

For read-only strings (i.e. embedded and with the compiler) and with
threads it makes no sense to share the volatile cowrefcnt with the
read-only buffer of the string. We cannot use read-only strings.  And
it's problematic to change the read-onlyness of the COW'ed string just
to bump the refcnt.

So move the CowREFCNT to the HEAD, to be shared with sv_refcnt.
Remove all the code which dealt with CowREFCNT at the end of the string.

We don't inc both refcounts above MAX, but we keep silent in refcnt, and
leave potential leaks.  In practice the highest observed refcnt was
118480, MAX is 16777216.  The cowrefnt is still cap'ed at 255.