inline: optimize args
check the body for all @_ args, and optionally
splice the args in, either by ref or as copy.
do this only for max 5 args.
check some arg indices at compile-time (asserts only)
TODO: we should change it to undef if the index is oob
or skip optim_arg then.

when the pushmark - entersub sequence has more
pushmarks at the beginning, start with the 2nd/last